Add factorial function Kak ida/factorial
